Issue-758: Moved PersonDetailsData to correspondent package
