+++ screen/person/src/main/java/com/kirchhoff/movies/screen/person/ui/screen/details/model/PersonDetailsScreenState.kt
@@ -1,7 +1,5 @@
 package com.kirchhoff.movies.screen.person.ui.screen.details.model
 
-import com.kirchhoff.movies.screen.person.data.UIPersonCredits
-import com.kirchhoff.movies.screen.person.data.UIPersonDetails
 import com.kirchhoff.movies.screen.person.data.UIPersonImage
 
 internal data class PersonDetailsScreenState(
